Finally figure out and finish `cpufreq.js`
No, this won't be replacing the `cpufreq` written in Perl, because this
isn't exactly ideal, although it's handy that it doesn't require the
use of `/proc/cpuinfo`.

My next goal may be to try to actually go with the aforementioned
file, in order to learn how to parse it the way I did in Perl.